Rename Human to Humanoid, Make Human both Humanoid and Creature.
With what may seem like an odd change is necessary because we desire to
share many things between Player and Human.

Humans in Sponge are no different than a Skeleton, Zombie, etc...they are
just "Creature"s. As a Creature, I can easily give them AI in a plugin
which means the obvious choice would be Human IS-A Creature. This presents
problems because Player would then be capable of AI and while that seems
silly...if the AI allows it, someone will attempt it and this would fail
epically.

Simple fix is a common interface between Human and Player and keep AI to
Human for Humanoid types.

Lastly, add the ability to remove tasks.

Signed-off-by: Chris Sanders <zidane@outlook.com>