Add BoundAITasks/AIBuilders for when a task has a known Owner.
BoundAITasks are tasks in-which we, for sure, know the owner because the
task is being created. This is a difference from AbstractAITask where, due
to it being wrapped in the impl, we don't know the owner until it being
added.

AIBuilders are used to allow the API to create Vanilla AI objects to
attach to entities. Currently, I've provided a builder for the WanderAI so
that we can see how this will work.

Signed-off-by: Chris Sanders <zidane@outlook.com>